Bash cli tooling added for install/setup

+++ install.sh
@@ -3,6 +3,14 @@
 
 set -o pipefail
 
+function _has_cmd() {
+  type $1 &>/dev/null
+}
+
+function _has_pkg(){
+  dpkg -s $1 &>/dev/null
+}
+
 function _mkcd(){
   mkdir -p $1 && cd $1
 }
@@ -26,30 +34,42 @@ function _install_gh_deb(){(
   local DL_PATH=${2:?}
   local DL_DEB=${DL_PATH##*/}
   _mkcd ~/Downloads
+  [ -f $DL_DEB ] && exit 0
   curl -OL https://github.com/$REPO/releases/download/$DL_PATH || exit $?
   sudo apt-fast install ./$DL_DEB
 )}
 
 function install-keychain(){
-  sudo apt-fast install -y keychain
+  echo $FUNCNAME
+  _has_cmd keychain || sudo apt-fast install -y keychain
 }
 
 function install-subrepo(){
-  git clone --filter=blob:none https://github.com/ingydotnet/git-subrepo ~/.local/share/git-subrepo
+  echo $FUNCNAME
+  local REPO='~/.local/share/git-subrepo'
+  if [ -d $REPO ]; then
+    cd $REPO && git pull
+  else
+    git clone --filter=blob:none https://github.com/ingydotnet/git-subrepo $REPO
+  fi
 }
 
 function install-git-delta(){
-  local ARCH VERSION REPO
+  echo $FUNCNAME
+  local ARCH VERSION REPO SHARE
   REPO="dandavison/delta"
   ARCH=$(_dist_arch) || return $?
   VERSION=$(_gh_latest_version $REPO) || return $?
+  SHARE='~/.local/share/delta/'
   _install_gh_deb $REPO "${VERSION}/git-delta_${VERSION}_${ARCH}.deb" || return $?
+  [ -f $SHARE/themes.gitconfig ] && return
   (
-  _mkcd ~/.local/share/delta
+  _mkcd $SHARE
   curl -sLO https://github.com/dandavison/delta/raw/refs/heads/main/themes.gitconfig
 )}
 
 function install-bat(){
+  echo $FUNCNAME
   local ARCH VERSION REPO
   REPO="sharkdp/bat"
   ARCH=$(_dist_arch) || return $?
@@ -58,32 +78,42 @@ function install-bat(){
 }
 
 function install-tmux(){
-  sudo apt-fast install -y tmux
+  echo $FUNCNAME
+  _has_cmd tmux || sudo apt-fast install -y tmux
 }
 
 function install-vim(){
-  sudo add-apt-repository -y ppa:jonathonf/vim && \
-  sudo apt-fast install -y vim-gtk3
+  echo $FUNCNAME
+  grep -qr 'jonathonf/vim' /etc/apt/sources.list* \
+  || { sudo add-apt-repository -y ppa:jonathonf/vim && sudo apt-fast update; }
+  _has_pkg vim-gtk3 || sudo apt-fast install -y vim-gtk3
 }
 
 function install-pipx(){
+  echo $FUNCNAME
   # System python is installed exclusively for installing pipx
+  _has_cmd pipx || { \
   sudo apt-fast install -y python3-pip && \
   python3 -m pip install --user pipx && \
-  python3 -m pipx ensurepath
+  python3 -m pipx ensurepath ; }
 }
 
 function install-common-apt(){
-  sudo apt-fast install -y silversearcher-ag
+  echo $FUNCNAME
+  _has_cmd ag || sudo apt-fast install -y silversearcher-ag
 }
 
 function install-bash-alias-completion(){
-  mkdir ~/.bash_completion.d
+  echo $FUNCNAME
+  [ -f ~/.bash_completion.d/complete_alias ] && return
+  mkdir -p ~/.bash_completion.d
   curl https://raw.githubusercontent.com/cykerway/complete-alias/master/complete_alias \
        > ~/.bash_completion.d/complete_alias
 }
 
 function install-docker(){
+  echo $FUNCNAME
+  _has_cmd docker && docker version &>/dev/null && return
   # Add Docker's official GPG key:
   sudo install -m 0755 -d /etc/apt/keyrings
   sudo curl -fsSL https://download.docker.com/linux/ubuntu/gpg -o /etc/apt/keyrings/docker.asc || return $?
@@ -109,7 +139,43 @@ function install-docker(){
   _install_gh_deb $REPO "v${VERSION}/dive_${VERSION}_linux_${ARCH}.deb"
 }
 
-# TODO: add the cli bash script here
+# Utilities
+# https://stackoverflow.com/questions/2683279/how-to-detect-if-a-script-is-being-sourced
+(return 0 2>/dev/null) && sourced=1 || sourced=0
+
+function _install-list(){
+declare -F | awk '$3~/^install-/{sub(/^install-/,"",$3); print $3}'
+}
+
+# If not sourced
+if [ "$sourced" -eq "0" ]; then
+  function _arg-alias(){ sed -r "s/(^| )$1($| )/\1$2 /g"; }
+  function _arg-alias-list-expr(){
+    printf -- 'sed -r '
+    _install-list | while read -r short; do
+      printf -- '-e "s/(^| )%s($| )/\\1%s /g" ' $short install-$short
+    done
+  }
+  ## Filters: pick one
+  # Multiple cmds chained with '--', optional leading '--' on first arg
+  # ./this.sh [--] fn1 [arg1 ...] [-- fn2] ...
+  function _chain_dashes(){ sed -r -e "s/^ *--//g" -e "s/--/ \&\& /g"; }
+  # Multiple cmds, taking no args, whitespace sep
+  # ./this.sh fn1 [fn2] ...
+  function _chain_space(){ sed -r -e "s/^ *//g" -e "s/ +/ \&\& /g"; }
+  ## Cmds:
+  if [ -z "$1" ]; then
+    _install-list
+  elif [ "$1" == "all" ]; then
+    _install-list | while read -r short; do
+      eval install-$short || { echo "ERROR $?"; break; }
+    done
+  else
+    ARGS=$(echo "$*" | _chain_space | eval $(_arg-alias-list-expr))
+    eval $ARGS || echo "ERROR $?"
+    echo Done
+  fi
+fi
 
 # install.sh
